Split routes in to separate files
